% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Avgs.R
\name{Calc_Avgs}
\alias{Calc_Avgs}
\title{Takes filtered results from TimePeriods.R and arithmetically averages all decibel levels to a single figure}
\usage{
Calc_Avgs(results, num_periods)
}
\arguments{
\item{results}{The filtered results from TimePeriods.R}

\item{num_periods}{The number of periods for analsysis in TimePeriods.R}
}
\value{
A new single row dataframe of the averaged decibel levels
}
\description{
Creates a new dataframe and csv file
}
\examples{
Calc_Avgs(results,num_periods = 2)
}
