% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_LogAvgs.R
\name{Calc_LogAvgs}
\alias{Calc_LogAvgs}
\title{Takes filtered results from TimePeriods.R and logaverages all decibel levels to a single figure}
\usage{
Calc_LogAvgs(results, num_periods)
}
\arguments{
\item{results}{The filtered results from TimePeriods.R}

\item{num_periods}{The number of periods for analsysis in TimePeriods.R}
}
\value{
A new single row dataframe of the log-averaged decibel levels
}
\description{
Creates a new dataframe and csv file
}
\examples{
Calc_LogAvgs(results,num_periods = 2)
}
