% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conv2Oct.R
\name{Conv2Oct}
\alias{Conv2Oct}
\title{Converts Svan 1/3 octave data to octave data}
\usage{
Conv2Oct(df)
}
\arguments{
\item{df}{the svan data in third octaves}

\item{center_freqs}{The octave band centre freqs}
}
\value{
A new dataframe in octave bands and exports a CSV to a folder
}
\description{
This function takes a csv data input, skips the first 14 cols
and log sums the third octave data to octave data
}
\examples{
Conv2Oct(file_path)
}
