% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_xlsx.R
\name{Create_xlsx}
\alias{Create_xlsx}
\title{Creates an xlsx from all the csvs in the CSV folder}
\usage{
Create_xlsx()
}
\arguments{
\item{csv_files}{Directory to folder holding all csv files}
}
\value{
Output excel file in XLSX folder
}
\description{
Creates an .xlsx file
}
\examples{
Create_xlsx()
}
