% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimePeriods.R
\name{TimePeriods}
\alias{TimePeriods}
\title{Filters octave data by period and target level}
\usage{
TimePeriods(data, num_periods)
}
\arguments{
\item{data}{is the octave band data as an output dataframe from Conv2Oct}

\item{num_periods}{number of periods for analysis}
}
\value{
New filtered dataframes (periods and target levels) in octave bands and exports a CSV to a folder
}
\description{
Choose number of periods for analysis, enter time periods
and enter target levels - levels within 0.5 dB of target level will be filtered for further analysus
}
\examples{
TimePeriods(oct_df,num_periods = 2)
}
